libefivar: avoid accessing type-punned pointer
Fix -Werror=strict-aliasing in Strtoi64() macro exposed when building
world with -O2.

+++ lib/libefivar/uefi-dplib.h
@@ -526,7 +526,7 @@ UefiDevicePathLibConvertTextToDevicePath (
 #define StrnCmp(a, b, n) strncmp(a, b, n)
 #define StrnLenS(str, max) strlen(str)
 #define Strtoi(x) strtol(x, NULL, 0)
-#define Strtoi64(x, y) *(long long *)y = strtoll(x, NULL, 0)
+#define Strtoi64(x, y) *(UINT64 *)y = strtoll(x, NULL, 0)
 #define SwapBytes64(u64) bswap64(u64)
 #define UnicodeStrToAsciiStrS(src, dest, len) strlcpy(dest, src, len)
 #define ZeroMem(p,l) memset(p, 0, l)
